requirements.txt
# Notion API
httpx>=0.27.0

# LangGraph + LLM
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0
langchain-core>=0.3.0

# PDF 텍스트 추출
pymupdf>=1.24.0

# Env
python-dotenv>=1.0.0

# Optional: Slack
# slack-sdk>=3.27.0
